Fix compiler warning, avoid overflow

+++ src/rprof_freelist.cpp
@@ -58,4 +58,4 @@ void rprofFreeListFree(struct rprofFreeList_t* _freeList, void* _ptr)
 		_freeList->m_next	= (uint8_t*)_ptr;
 	}
 	++_freeList->m_blocksFree;
-}int rprofFreeListCheckPtr(struct rprofFreeList_t* _freeList, void* _ptr){	return	(uintptr_t)(_freeList->m_maxBlocks * _freeList->m_blockSize) > 			(uintptr_t)(((uint8_t*)_ptr) - _freeList->m_buffer) ? 1 : 0;}
+}int rprofFreeListCheckPtr(struct rprofFreeList_t* _freeList, void* _ptr){	return	((uintptr_t)_freeList->m_maxBlocks * (uintptr_t)_freeList->m_blockSize) > 			(uintptr_t)(((uint8_t*)_ptr) - _freeList->m_buffer) ? 1 : 0;}
